# screen_system/openapi_before.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ShipIn": {
        "properties": {
          "bizlogi_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bizlogi Status"
          },
          "order_no": {
            "title": "Order No",
            "type": "string"
          },
          "package_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Package Count"
          },
          "shipping_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Shipping Status"
          },
          "tracking_no": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tracking No"
          }
        },
        "required": [
          "order_no"
        ],
        "title": "ShipIn",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "スクリーン原団 工程管理 v2",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/admin/prints": {
      "get": {
        "operationId": "admin_prints_api_admin_prints_get",
        "parameters": [
          {
            "in": "header",
            "name": "x-admin-pass",
            "required": false,
            "schema": {
              "default": "",
              "title": "X-Admin-Pass",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Admin Prints"
      }
    },
...
    "/api/admin/shipments": {
      "get": {
        "operationId": "admin_shipments_api_admin_shipments_get",
        "parameters": [
          {
            "in": "header",
            "name": "x-admin-pass",
            "required": false,
            "schema": {
              "default": "",
              "title": "X-Admin-Pass",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Admin Shipments"
      },
      "post": {
        "operationId": "admin_shipment_upsert_api_admin_shipments_post",
        "parameters": [
          {
            "in": "header",
            "name": "x-admin-pass",
            "required": false,
            "schema": {
              "default": "",
              "title": "X-Admin-Pass",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ShipIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Admin Shipment Upsert"
      }
    },
...
    "/api/admin/sync-logs": {
      "get": {
        "operationId": "admin_sync_logs_api_admin_sync_logs_get",
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "default": "failed",
              "title": "Status",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-admin-pass",
            "required": false,
            "schema": {
              "default": "",
              "title": "X-Admin-Pass",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Admin Sync Logs"
      }
    },
...
  }
}